rename FunctionReplaceRector to RenameFunctionRector
