+++ config/level/php/php72.yaml
@@ -8,7 +8,7 @@ services:
     Rector\Php\Rector\Unset_\UnsetCastRector: ~
     Rector\Php\Rector\ConstFetch\BarewordStringRector: ~
 
-    Rector\Rector\Function_\FunctionReplaceRector:
+    Rector\Rector\Function_\RenameFunctionRector:
         jpeg2wbmp: 'imagecreatefromjpeg' # and imagewbmp
         png2wbmp: 'imagecreatefrompng' # or imagewbmp
         # http://php.net/manual/en/migration72.deprecated.php#migration72.deprecated.gmp_random-function
