+++ config/level/php/php73.yaml
@@ -6,7 +6,7 @@ services:
     Rector\Php\Rector\String_\SensitiveHereNowDocRector: ~
 
     # https://wiki.php.net/rfc/deprecations_php_7_3
-    Rector\Rector\Function_\FunctionReplaceRector:
+    Rector\Rector\Function_\RenameFunctionRector:
         image2wbmp: 'imagewbmp'
         mbregex_encoding: 'mb_regex_encoding'
         mbereg: 'mb_ereg'
