+++ config/level/php/php74.yaml
@@ -1,7 +1,7 @@
 services:
     Rector\Php\Rector\Property\TypedPropertyRector: ~
 
-    Rector\Rector\Function_\FunctionReplaceRector:
+    Rector\Rector\Function_\RenameFunctionRector:
         # https://wiki.php.net/rfc/deprecations_php_7_4#the_real_type
         is_float: 'is_real'
         # https://wiki.php.net/rfc/deprecations_php_7_4#apache_request_headers_function
